backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.8
pydantic<2.0.0,>=1.8.2
uvicorn==0.24.0
pyjwt==2.8.0
python-multipart==0.0.6
python-dotenv==1.0.0
SQLAlchemy==1.4.41
psycopg2-binary==2.9.11